+++ src/egon/data/datasets/heat_supply/__init__.py
@@ -113,6 +113,6 @@ class HeatSupply(Dataset):
             name="HeatSupply",
             version="0.0.0",
             dependencies=dependencies,
-            tasks=(create_tables, {
-                district_heating, individual_heating, potential_germany}),
+            tasks=(create_tables,
+                district_heating, individual_heating, potential_germany),
         )